+++ java/org/apache/el/lang/ELArithmetic.java
@@ -64,7 +64,7 @@ public abstract class ELArithmetic {
 
         @Override
         protected Number mod(Number num0, Number num1) {
-            return Double.valueOf(num0.doubleValue() % num1.doubleValue());
+            return num0.doubleValue() % num1.doubleValue();
         }
 
         @Override
@@ -133,7 +133,7 @@ public abstract class ELArithmetic {
             } else if (num1 instanceof BigDecimal) {
                 return ((new BigDecimal(num0.doubleValue()).add((BigDecimal) num1)));
             }
-            return Double.valueOf(num0.doubleValue() + num1.doubleValue());
+            return num0.doubleValue() + num1.doubleValue();
         }
 
         @Override
@@ -142,7 +142,7 @@ public abstract class ELArithmetic {
                 return num;
             if (num instanceof BigInteger)
                 return new BigDecimal((BigInteger) num);
-            return Double.valueOf(num.doubleValue());
+            return num.doubleValue();
         }
 
         @Override
@@ -152,12 +152,12 @@ public abstract class ELArithmetic {
 
         @Override
         protected Number divide(Number num0, Number num1) {
-            return Double.valueOf(num0.doubleValue() / num1.doubleValue());
+            return num0.doubleValue() / num1.doubleValue();
         }
 
         @Override
         protected Number mod(Number num0, Number num1) {
-            return Double.valueOf(num0.doubleValue() % num1.doubleValue());
+            return num0.doubleValue() % num1.doubleValue();
         }
 
         @Override
@@ -168,7 +168,7 @@ public abstract class ELArithmetic {
             } else if (num1 instanceof BigDecimal) {
                 return ((new BigDecimal(num0.doubleValue()).subtract((BigDecimal) num1)));
             }
-            return Double.valueOf(num0.doubleValue() - num1.doubleValue());
+            return num0.doubleValue() - num1.doubleValue();
         }
 
         @Override
@@ -179,7 +179,7 @@ public abstract class ELArithmetic {
             } else if (num1 instanceof BigDecimal) {
                 return ((new BigDecimal(num0.doubleValue()).multiply((BigDecimal) num1)));
             }
-            return Double.valueOf(num0.doubleValue() * num1.doubleValue());
+            return num0.doubleValue() * num1.doubleValue();
         }
 
         @Override
@@ -198,14 +198,14 @@ public abstract class ELArithmetic {
 
         @Override
         protected Number add(Number num0, Number num1) {
-            return Long.valueOf(num0.longValue() + num1.longValue());
+            return num0.longValue() + num1.longValue();
         }
 
         @Override
         protected Number coerce(Number num) {
             if (num instanceof Long)
                 return num;
-            return Long.valueOf(num.longValue());
+            return num.longValue();
         }
 
         @Override
@@ -215,22 +215,22 @@ public abstract class ELArithmetic {
 
         @Override
         protected Number divide(Number num0, Number num1) {
-            return Long.valueOf(num0.longValue() / num1.longValue());
+            return num0.longValue() / num1.longValue();
         }
 
         @Override
         protected Number mod(Number num0, Number num1) {
-            return Long.valueOf(num0.longValue() % num1.longValue());
+            return num0.longValue() % num1.longValue();
         }
 
         @Override
         protected Number subtract(Number num0, Number num1) {
-            return Long.valueOf(num0.longValue() - num1.longValue());
+            return num0.longValue() - num1.longValue();
         }
 
         @Override
         protected Number multiply(Number num0, Number num1) {
-            return Long.valueOf(num0.longValue() * num1.longValue());
+            return num0.longValue() * num1.longValue();
         }
 
         @Override
@@ -247,12 +247,12 @@ public abstract class ELArithmetic {
 
     public static final LongDelegate LONG = new LongDelegate();
 
-    private static final Long ZERO = Long.valueOf(0);
+    private static final Long ZERO = 0L;
 
     public static final Number add(final Object obj0, final Object obj1) {
         final ELArithmetic delegate = findDelegate(obj0, obj1);
         if (delegate == null) {
-            return Long.valueOf(0);
+            return 0L;
         }
 
         Number num0 = delegate.coerce(obj0);
@@ -263,7 +263,7 @@ public abstract class ELArithmetic {
 
     public static final Number mod(final Object obj0, final Object obj1) {
         if (obj0 == null && obj1 == null) {
-            return Long.valueOf(0);
+            return 0L;
         }
 
         final ELArithmetic delegate;
@@ -285,7 +285,7 @@ public abstract class ELArithmetic {
     public static final Number subtract(final Object obj0, final Object obj1) {
         final ELArithmetic delegate = findDelegate(obj0, obj1);
         if (delegate == null) {
-            return Long.valueOf(0);
+            return 0L;
         }
 
         Number num0 = delegate.coerce(obj0);
@@ -316,7 +316,7 @@ public abstract class ELArithmetic {
     public static final Number multiply(final Object obj0, final Object obj1) {
         final ELArithmetic delegate = findDelegate(obj0, obj1);
         if (delegate == null) {
-            return Long.valueOf(0);
+            return 0L;
         }
 
         Number num0 = delegate.coerce(obj0);
@@ -385,7 +385,7 @@ public abstract class ELArithmetic {
             return coerce((String) obj);
         }
         if (obj instanceof Character) {
-            return coerce(Short.valueOf((short) ((Character) obj).charValue()));
+            return coerce((short) ((Character) obj).charValue());
         }
 
         throw new IllegalArgumentException(MessageFactory.get("error.convert",
